requirements.txt
PySimpleGUI
lupa
parsimonious
Pillow
Jinja2